Code reorganization without functional changes. Consolidated Node, moved the file operation out of the decider (passing TM instead), split in library and executable and added a benchmark.
