+++ decider-halting-segment-reproduction/decider_halting_segment/src/node.rs
@@ -1,6 +1,59 @@
-use crate::*;
+use std::fmt;
+
+use crate::{tm::HeadMove, *};
+
+pub(crate) enum NodeLimit {
+    NoLimit,
+    NodeLimit(usize),
+}
+
+#[derive(Copy, Clone, PartialEq, Eq, Hash)]
+pub enum OutsideSegmentOrState {
+    OutsideSegment,
+    State(u8),
+}
+
+#[derive(Copy, Clone, PartialEq, Eq, Hash)]
+pub enum SegmentCell {
+    Unallocated,
+    Bit(bool),
+}
+
+#[derive(Clone, PartialEq, Eq, Hash)]
+pub(crate) struct SegmentCells(pub Vec<SegmentCell>);
+
+impl SegmentCells {
+    fn are_there_no_ones(&self) -> bool {
+        !self
+            .0
+            .iter()
+            .any(|cell| matches!(cell, SegmentCell::Bit(true)))
+    }
+}
+
+#[derive(Clone, PartialEq, Eq, Hash)]
+pub(crate) struct Node {
+    pub(crate) state: OutsideSegmentOrState,
+    pub(crate) segment: SegmentCells,
+    pub(crate) pos_in_segment: usize,
+}
 
 impl Node {
+    pub(crate) fn is_fatal(&self) -> bool {
+        /* Fatal nodes are nodes whose segment contain no 1s and:
+           - head is outside segment
+           - Or, state is A
+        Detecting these nodes is important because when the decider meets one, we know
+        that we cannot conclude that the machine does not halt.
+        */
+        match self.state {
+            OutsideSegmentOrState::OutsideSegment | OutsideSegmentOrState::State(0) => {
+                self.segment.are_there_no_ones()
+            }
+            _ => false,
+        }
+    }
+
     pub fn get_neighbours(&self, tm: &TM) -> Vec<Node> {
         /* Returns the halting segment neighbours of the node:
 
@@ -143,5 +196,90 @@ impl Node {
         }
 
         to_return
+    }
+}
+
+impl fmt::Display for Node {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        /*
+            State: E ;  _ . .[0]. . _  ;
+        */
+        let state_char: char = match self.state {
+            OutsideSegmentOrState::OutsideSegment => '*',
+            OutsideSegmentOrState::State(state) => (b'A' + state) as char,
+        };
+
+        write!(f, "State: {} ; ", state_char)?;
+
+        let is_outside = matches!(self.state, OutsideSegmentOrState::OutsideSegment);
+
+        if is_outside && self.pos_in_segment == 0 {
+            write!(f, "[_]")?;
+        } else {
+            write!(f, " _")?;
+        }
+
+        for i in 0..self.segment.0.len() {
+            if i == 0 && (i != self.pos_in_segment && !is_outside) {
+                write!(f, " ")?;
+            }
+
+            let mut space_after = true;
+            match self.segment.0[i] {
+                SegmentCell::Unallocated => write!(f, ".")?,
+                SegmentCell::Bit(bit) => {
+                    if i != self.pos_in_segment || is_outside {
+                        write!(f, "{}", bit as u8)?;
+                    } else {
+                        write!(f, "[{}]", bit as u8)?;
+                        space_after = false;
+                    }
+                }
+            }
+            if space_after && (i + 1 != self.pos_in_segment || is_outside) {
+                write!(f, " ")?;
+            }
+        }
+
+        if is_outside && self.pos_in_segment + 1 == self.segment.0.len() {
+            write!(f, "[_]")?;
+        } else {
+            write!(f, " _")?;
+        }
+
+        write!(f, "")
+    }
+}
+
+// pub(crate) struct Nodes(pub Vec<Node>);
+//
+// impl fmt::Display for Nodes {
+//     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+//         for (i, node) in self.0.iter().enumerate() {
+//             writeln!(f, "{}: {}", i, node)?;
+//         }
+//         write!(f, "")
+//     }
+// }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn machine_324_trace() {
+        let chaotic_machine_id = 324;
+        let tm: TM = TM::from_bbchallenge_id(chaotic_machine_id, PATH_TO_BBCHALLENGE_DB_TEST).unwrap();
+
+        halting_segment_decider(&tm, 5, 2, NodeLimit::NodeLimit(1000), true);
+    }
+
+    #[test]
+    fn machine_chaotic_trace() {
+        // Chaotic Machine [Marxen & Buntrock, 1990]
+        let chaotic_machine_id = 76708232;
+        let tm: TM = TM::from_bbchallenge_id(chaotic_machine_id, PATH_TO_BBCHALLENGE_DB_TEST).unwrap();
+
+        halting_segment_decider(&tm, 5, 2, NodeLimit::NodeLimit(1000), true);
     }
 }
