+++ decider-halting-segment-reproduction/decider_halting_segment/src/utils.rs
